dscicd/requirements.txt
anthropic==0.34.1
boto3==1.35.9
dnspython==2.6.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.6
marshmallow==3.22.0
pymongo==3.10.1
PyPDF2==3.0.1
python-dotenv==1.0.1
reportlab==4.2.2
requests==2.31.0
flask[async]
aiohttp